refactor event tracking hook
