+++ src/app/containers/CpsTopStories/EventTrackingContext.js
@@ -1,0 +1,3 @@
+import { createContext } from 'react';
+
+export default createContext({});
